Refactor: modulariza lógica de regiones temporales y listeners
