+++ src/main/java/com/xiluiis/temporaryprotections/ConfigManager.java
@@ -1,0 +1,5 @@
+package com.xiluiis.temporaryprotections;
+
+public class ConfigManager {
+    
+}
