+++ src/main/java/com/xiluiis/temporaryprotections/DebugManager.java
@@ -1,0 +1,5 @@
+package com.xiluiis.temporaryprotections;
+
+public class DebugManager {
+    
+}
