+++ src/main/resources/config.yml
@@ -1,0 +1,3 @@
+# config.yml
+temporary-protection-seconds: 60
+debug-messages: true
